+++ QGearsMain/src/core/XmlBackground2DFile.cpp
@@ -68,7 +68,7 @@ XmlBackground2DFile::Load()
                 Ogre::Vector4 distance = Ogre::Vector4( 0, 0, -depth, 1 );
                 Ogre::Vector4 res = CameraManager::getSingleton().GetCurrentCamera()->getProjectionMatrixWithRSDepth() * distance;
                 res = res / res.w;
-                background->AddTile( destination.x, destination.y, width, height, res.z, uv.x, uv.y, uv.z, uv.w, blending );
+                background->AddTile( static_cast<int>(destination.x), static_cast<int>(destination.y), width, height, res.z, uv.x, uv.y, uv.z, uv.w, blending );
 
                 TiXmlNode* node2 = node->FirstChild();
                 while( node2 != NULL )
